+++ PythonDecoratorSkipRuntime/log/SqlDecoratorSkipCheck_log.txt
@@ -1,0 +1,28 @@
+OUTPUT: 
+  Note:: Please ignore sql server connection error
+
+    SqlDecoratorSkipCheck class init
+    From decorator function
+    Function name is: <function SqlDecoratorSkipCheck.test at 0x000001C2916BBB70>
+    Creating new object here!
+    Observed exception: ('08001', '[08001] [Microsoft][ODBC SQL Server Driver][DBNETLIB]SQL Server does not exist or access denied. (17) (SQLDriverConnect); [08001] [Microsoft][ODBC SQL Ser
+    ver Driver][DBNETLIB]ConnectionOpen (Connect()). (2)')
+    self dict: {'fail': 'FAIL', 'log_obj': None, 'user_lib_obj': None, 'cursor_obj': 'HARDCODE value'}
+    kwargs: {}
+    cursor_obj is already created|updated: HARDCODE value
+    self dict: {'fail': 'FAIL', 'log_obj': None, 'user_lib_obj': None, 'cursor_obj': 'HARDCODE value'}
+    kwargs: {}
+    cursor_obj is already created|updated: HARDCODE value
+    self dict: {'fail': 'FAIL', 'log_obj': None, 'user_lib_obj': None, 'cursor_obj': None}
+    kwargs: {}
+    From decorator function
+    Function name is: <function SqlDecoratorSkipCheck.test_4 at 0x000001C2916C00D0>
+    Creating new object here!
+    Observed exception: ('08001', '[08001] [Microsoft][ODBC SQL Server Driver][DBNETLIB]SQL Server does not exist or access denied. (17) (SQLDriverConnect); [08001] [Microsoft][ODBC SQL Ser
+    ver Driver][DBNETLIB]ConnectionOpen (Connect()). (2)')
+    self dict: {'fail': 'FAIL', 'log_obj': None, 'user_lib_obj': None, 'cursor_obj': 'HARDCODE value'}
+    kwargs: {}
+    self dict: {'fail': 'FAIL', 'log_obj': None, 'user_lib_obj': None, 'cursor_obj': 'HARDCODE value'}
+    kwargs: {'skip_decorator': True, 'muthu': 123}
+    I am from class method
+    I am from static method
