+++ PythonRandomStrGen/log/RamdomStrGenerate_log.txt
