--- Python_random_ipv4_ipv6_gen/log/RandomIpaddress_log.txt
@@ -1,10 +1,0 @@
-OUTPUT:
-	ipv_6_and_ipv4: {'ipv4': '151.80.207.222', 'ipv6': 'ec8c:becb:1c5f:e22e:9090:961d:aa0d:18ca'}
-	ipv_6: {'ipv6': 'a1ea:13fe:d1a3:7ada:8928:561d:2642:97ae'}
-	ipv_6: {'ipv6': '32f2:a1cd:a92e:d2bf:b46e:aeb3:fadd:bbfc'}
-	ipv_6: {'ipv6': '419a:fcff:fbfd:dae5:b821:1dfd:781e:f9ec'}
-	ipv_4: {'ipv4': '47.19.53.69'}
-	ipv_4: {'ipv4': '40.16.86.21'}
-	ipv_4: {'ipv4': '238.55.145.155'}
-	Default ip-address type: {'ipv4': '91.48.198.175'}
-	Invalid ipaddress: {}
