+++ README.md
@@ -18,3 +18,5 @@
   * ***Polymorphism with Inheritance***
   * ***Polymorphism with a Function and objects***
 * **Inheritance**
+* **ExcelSheetWrite**
+* **DateDifferenceCalculateBeforeOrAfter**
